Remove deprecated JDBC class and move EncodingType to azkaban.db (#1414)
* remove Abstract JDBC class

This code patch does two things:
* Remove Deprecated Abstract JDBC class
* move EncodingType class to azkaban.db module
