+++ azkaban-common/src/main/java/azkaban/executor/FetchActiveFlowDao.java
@@ -16,7 +16,7 @@
 
 package azkaban.executor;
 
-import azkaban.database.EncodingType;
+import azkaban.db.EncodingType;
 import azkaban.db.DatabaseOperator;
 import azkaban.utils.GZIPUtils;
 import azkaban.utils.Pair;
